split up project and fixture modules
