+++ src/all_fixtures.json
@@ -24,7 +24,8 @@
         },
         {
           "channel": 5,
-          "label": "Dimmer"
+          "label": "Dimmer",
+          "home": 0
         },
         {
           "channel": 6,
@@ -90,9 +91,25 @@
           "channel": 11,
           "label": "CTO",
           "notes": "Colour Temperature Orange"
+        },
+        {
+          "channel": 17,
+          "label": "Zoom",
+          "notes": "Narrow to wide"
+        },
+        {
+          "channel": 18,
+          "label": "Focus",
+          "notes": "0% to 100% "
         }
       ],
       "macros": [
+        {
+          "control": {
+            "label": "brightness",
+            "channels": [5]
+          }
+        },
         {
           "colour": {
             "label": "colour",
@@ -116,6 +133,18 @@
           "control": {
             "label": "tilt",
             "channels": [3]
+          }
+        },
+        {
+          "control": {
+            "label": "zoom",
+            "channels": [17]
+          }
+        },
+        {
+          "control": {
+            "label": "focus",
+            "channels": [18]
           }
         }
       ]
@@ -153,7 +182,7 @@
           "channel": 5,
           "label": "Intensity coarse",
           "notes": "Intensity 0 -> 100%",
-          "home": 255
+          "home": 0
         },
         {
           "channel": 6,
@@ -709,30 +738,43 @@
           ]
         }
       ],
-      "macros": [
+      "macros": [ 
         {
-          "label": "red",
-          "channels": [12]
+          "control": {
+            "label": "brightness",
+            "channels": [5]
+          }
         },
         {
-          "label": "green",
-          "channels": [14]
+          "colour": {
+            "label": "colour",
+            "channels": {
+              "additive": {
+                "red": [12],
+                "green": [14],
+                "blue": [16],
+                "white": [18]
+              }
+            }
+          }
         },
         {
-          "label": "blue",
-          "channels": [16]
+          "control": {
+            "label": "pan",
+            "channels": [1]
+          }
         },
         {
-          "label": "pan",
-          "channels": [1]
+          "control": {
+            "label": "tilt",
+            "channels": [3]
+          }
         },
         {
-          "label": "tilt",
-          "channels": [3]
-        },
-        {
-          "label": "zoom",
-          "channels": [8]
+          "control": {
+            "label": "zoom",
+            "channels": [8]
+          }
         }
       ]
     }
@@ -1121,7 +1163,7 @@
         },
         {
           "channel": 3,
-          "label": "Inclination (tilt) ",
+          "label": "Inclination (tilt)",
           "notes": "(0° up to the maximum value of the Tilt area. middle position: 128)",
           "home": 128
         },
